+++ python/ambassador/ir/irtcpmapping.py
@@ -5,6 +5,7 @@ from ..config import Config
 
 from .irbasemapping import IRBaseMapping
 from .irbasemappinggroup import IRBaseMappingGroup
+from .irhttpmapping import IRHTTPMapping
 from .irtcpmappinggroup import IRTCPMappingGroup
 
 import hashlib
@@ -22,6 +23,7 @@ class IRTCPMapping (IRBaseMapping):
 
     AllowedKeys: ClassVar[Dict[str, bool]] = {
         "address": True,
+        "circuit_breakers": False,
         "enable_ipv4": True,
         "enable_ipv6": True,
         "host": True,
@@ -67,6 +69,20 @@ class IRTCPMapping (IRBaseMapping):
     @staticmethod
     def group_class() -> Type[IRBaseMappingGroup]:
         return IRTCPMappingGroup
+
+    def setup(self, ir: 'IR', aconf: Config) -> bool:
+        if not super().setup(ir, aconf):
+            return False
+
+        if self.get('circuit_breakers', None) is None:
+            self['circuit_breakers'] = ir.ambassador_module.circuit_breakers
+
+        if self.get('circuit_breakers', None) is not None:
+            if not IRHTTPMapping.validate_circuit_breakers(ir, self['circuit_breakers']):
+                self.post_error("Invalid circuit_breakers specified: {}, invalidating mapping".format(self['circuit_breakers']))
+                return False
+
+        return True
 
     def bind_to(self) -> str:
         bind_addr = self.get('address') or '0.0.0.0'